+++ c/dsls/lemon_fsm.h
@@ -1,0 +1,14 @@
+#ifndef LEMON_FSM_H
+#define LEMON_FSM_H
+
+#include "lemon_structs.h"
+
+// xxx protoize
+void FindRulePrecedences();
+void FindFirstSets();
+void FindStates();
+void FindLinks();
+void FindFollowSets();
+void FindActions();
+
+#endif // LEMON_FSM_H
